app/requirements.txt
alembic==1.4.2
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
idna==2.9
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.8
wcwidth==0.1.9
yapf==0.29.0
